requirements.txt
openai
google-generativeai
pandas
tqdm
python-dotenv
requests
